update worker script, add run scripts for worker server and redis

+++ system_server/worker_scripts/retrieve_models.py
@@ -5,9 +5,15 @@ import zipfile, io
 import base64
 import io
 import time
+import uuid
 from collections import defaultdict
 from io import StringIO
 from io import BytesIO
+from pymongo import MongoClient
+
+client            = MongoClient("172.18.0.1",authMechanism='SCRAM-SHA-256')
+job_collection    = client["fvonprem"]["jobs"]
+models_collection = client["fvonprem"]["models"]
 
 def base_path():
     xavier_ssd = '/xavier_ssd/'
@@ -18,23 +24,22 @@ BASE_PATH_TO_MODELS = base_path()+'models/'
 def create_config_file(data):
     with open (BASE_PATH_TO_MODELS+'model.config', 'a') as f:
         f.write('model_config_list {\n')
-        for model in data:
-            if len(model['versions']) > 0:
-                f.write('\tconfig {\n')
-                f.write('\t\tname: \''+model['type']+'\'\n')
-                f.write('\t\tbase_path: \''+'/models/'+model['type']+'/\'\n')
-                f.write('\t\tmodel_platform: \'tensorflow\'\n')
-                f.write('\t\tmodel_version_policy: {all {}}\n')
-                f.write('\t}\n')
+        for model_data in data:
+            f.write('\tconfig {\n')
+            f.write('\t\tname: \''+model_data['type']+'\'\n')
+            f.write('\t\tbase_path: \''+'/models/'+model_data['type']+'/\'\n')
+            f.write('\t\tmodel_platform: \'tensorflow\'\n')
+            f.write('\t\tmodel_version_policy: {all {}}\n')
+            f.write('\t}\n')
         f.write('}')
 
-def retrieve_models(data):
+def retrieve_models(data, token):
+    job_id = str(uuid.uuid4())
+    insert_job_ref(job_id)
     if os.path.exists(BASE_PATH_TO_MODELS): os.system("rm -rf "+BASE_PATH_TO_MODELS)
     os.system("mkdir "+BASE_PATH_TO_MODELS)
 
-    #first create the config file from the models data
-    create_config_file(data)
-
+    models_versions = []
     for model_ref in data:
         project_id   = model_ref['_id']
         model_name   = model_ref['type']
@@ -42,39 +47,55 @@ def retrieve_models(data):
         model_folder = BASE_PATH_TO_MODELS + model_name
         if len(versions) > 0: os.system("mkdir " + model_folder)
 
-        #second iterate over the models data and request/extract model to models folder
+        model_data = {'type': model_name, 'versions': []}
+        #iterate over the models data and request/extract model to models folder
         for version in versions:
-
-            path = 'http://104.154.128.121/api/capture/models/download/'+str(project_id)+'/'+str(version)
-            #path = 'http://104.154.128.121/api/capture/models/download/c1948ab7-4e7c-48d1-810f-969328c90c23/1570413180585'
-
-            token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6IlJqTTRRa1kyTURnd1JrRTBNREZCT1VKR01UVkRNa0ZFTXpNMk9EVXdPRGxGUkRWRFJqZEVNZyJ9.eyJpc3MiOiJodHRwczovL2ZsZXhpYmxldmlzaW9uLmF1dGgwLmNvbS8iLCJzdWIiOiJhdXRoMHw1ZDlhMjMyNTQ5ZGIxYzBjNjM2Y2YwY2YiLCJhdWQiOiJodHRwczovL2ZsZXhpYmxldmlzaW9uL2FwaSIsImlhdCI6MTU3MDYzOTAyMiwiZXhwIjoxNTcwNzI1NDIyLCJhenAiOiI0dHpqWk5qRnl0RHE5S3FpRWg2MFpweVZ4UHF5dkRoSCIsInBlcm1pc3Npb25zIjpbXX0.sbqIFGkiWK1o3sBbwUccIy9oq_2WYy-LP4H-h3zsFeqAYg1dCDYBivhl52fx-3hRrcLwnPmKFelyR656HqMbbEdlqjmXSmm1UMabayeDxgCh0HLENGo7bBYWe3CXkYJr28Qroh5kSIJkyc7bRZIe8WPRnwsKkFb15QIrpkGTAAoErQAk1vP8GgwvjdbOZLOX90PxpumnYTBHE1W42bDQr1ocuQnKZXcKvkqc-grm6R5xJuglXRR6heLWY_csSlNxDjaDW2AhR9haOb9sqwV18ZxZhB-39wzmCr-Zsj5GjPmtd7hQo_bpwVgIEHqqtT2oxb1RBQaZlr7ZYV2JRFz-Vg"
-
+            path = 'http://104.154.128.121/api/capture/models/download/'+str(project_id)+'/'+str(version) 
             res = os.system(f"curl -X GET {path} -H 'accept: application/json' -H 'Authorization: Bearer {token}' -o {model_folder}/model.zip")
-            
-            print(res)
-            print('-----model download response-------------')
-            if res: print('yes we have a response')
-            
             try: 
                 with zipfile.ZipFile(model_folder+'/model.zip') as zf:
                     zf.extractall(model_folder)
-                
-                os.system("ls "+model_folder)
                 os.system("mv "+model_folder+"/job.json "+model_folder+"/"+str(version))
                 os.system("mv "+model_folder+"/object-detection.pbtxt "+model_folder+"/"+str(version))
+                model_data['versions'].append(version)
             except zipfile.BadZipfile:
                 print('bad zipfile in '+model_folder)
 
             os.system("rm -rf "+model_folder+'/model.zip')
-            os.system("ls "+model_folder)
 
-    print('removing old models')
-    os.system("docker exec localprediction rm -rf /models")
-    print('pushing new models to localprediction')
-    os.system("docker cp "+base_path()+"models localprediction:/")
-    return True
+        if model_data['versions']: models_versions.append(model_data)
+    
+    if models_versions:
+        create_config_file(models_versions)
+        save_models_versions(models_versions)
+        print('removing old models')
+        os.system("docker exec localprediction rm -rf /models")
+        print('pushing new models to localprediction')
+        os.system("docker cp "+base_path()+"models localprediction:/")
+        delete_job_ref(job_id)
+        return True
+    else: 
+        failed_job_ref(job_id)
+        return False
         
 
+def save_models_versions(models_versions):
+    models_collection.drop()
+    models_collection.insertMany(models_versions)
+
+def insert_job_ref(job_id):
+    job_collection.insert({
+        '_id': job_id,
+        'type': 'model_download',
+        'start_time': str(datetime.datetime.now()),
+        'status': 'downloading'
+        })
+
+def delete_job_ref(job_id):
+    query = {'_id': job_id}
+    job_collection.deleteOne(query)
 
 
+def failed_job_ref(job_id):
+    query = {'_id': job_id}
+    job_collection.updateOne(query, {$set: {'status': 'failed'}})
